requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
autoflake==2.3.1
black==25.1.0
cffi==1.17.1
click==8.2.1
colorlog==6.9.0
discord-protos==0.0.2
discord.py-self==2.0.1
frozenlist==1.6.0
idna==3.10
isort==6.0.1
multidict==6.4.4
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
PyNaCl==1.5.0
PyYAML==6.0.2
types-protobuf==6.30.2.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250515
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
